# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/api/v1/stock/summaries": {
            "get": {
                "description": "Find stock summaries by stock code, start date, and end date",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Stock"
                ],
                "summary": "Find stock summaries",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Stock code",
                        "name": "stock_code",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Start date (yyyy-mm-dd)",
                        "name": "start_date",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "End date (yyyy-mm-dd)",
                        "name": "end_date",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/model.StockSummaryResponse"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.Error"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.Error"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "model.StockSummaryResponse": {
            "type": "object",
            "properties": {
                "bid": {
                    "type": "number"
                },
                "bid_volume": {
                    "type": "number"
                },
                "change": {
                    "type": "number"
                },
                "close": {
                    "type": "number"
                },
                "date": {
                    "type": "string"
                },
                "delisting_date": {
                    "type": "string"
                },
                "first_trade": {
                    "type": "number"
                },
                "foreign_buy": {
                    "type": "number"
                },
                "foreign_sell": {
                    "type": "number"
                },
                "frequency": {
                    "type": "number"
                },
                "high": {
                    "type": "number"
                },
                "id_stock_summary": {
                    "type": "integer"
                },
                "index_individual": {
                    "type": "number"
                },
                "listed_shares": {
                    "type": "number"
                },
                "low": {
                    "type": "number"
                },
                "non_regular_frequency": {
                    "type": "number"
                },
                "non_regular_value": {
                    "type": "number"
                },
                "non_regular_volume": {
                    "type": "number"
                },
                "offer": {
                    "type": "number"
                },
                "offer_volume": {
                    "type": "number"
                },
                "open_price": {
                    "type": "number"
                },
                "percentage": {},
                "persen": {},
                "previous": {
                    "type": "number"
                },
                "remarks": {
                    "type": "string"
                },
                "stock_code": {
                    "type": "string"
                },
                "stock_name": {
                    "type": "string"
                },
                "tradeble_shares": {
                    "type": "number"
                },
                "value": {
                    "type": "number"
                },
                "volume": {
                    "type": "number"
                },
                "weight_for_index": {
                    "type": "number"
                }
            }
        },
...
        "response.Error": {
            "type": "object",
            "properties": {
                "field": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                }
            }
        }
    }
}